@@ -1,82 +1,35 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-	<Version>3.1.1</Version>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
-    <UserSecretsId>8d243b16-78f2-4542-8f36-682411560b1a</UserSecretsId>
+    <Version>4.2.0</Version>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <UserSecretsId>eca63fb9-b7f9-454f-910b-5088ae877085</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerfileContext>.</DockerfileContext>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <NoWarn>1591;1701;1702</NoWarn>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <NoWarn>1591;1701;1702</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="VersionNotes\**" />
-    <Content Remove="VersionNotes\**" />
-    <EmbeddedResource Remove="VersionNotes\**" />
-    <None Remove="VersionNotes\**" />
+    <Content Remove="Components\General\NavMenu.razor" />
   </ItemGroup>
 
   <ItemGroup>
-    <Content Remove="appsettings.Development.json" />
-    <Content Remove="appsettings.json" />
-    <Content Remove="settings\appsettings.Development.json" />
-    <Content Remove="settings\appsettings.json" />
-    <Content Remove="settings\namePolicyDefinition.json" />
-    <Content Remove="settings\resourcecomponents.json" />
-    <Content Remove="settings\resourcedelimiters.json" />
-    <Content Remove="settings\resourceenvironments.json" />
-    <Content Remove="settings\resourcefunctions.json" />
-    <Content Remove="settings\resourcelocations.json" />
-    <Content Remove="settings\generatednames.json" />
-	<Content Remove="settings\adminlog.json" />
-    <Content Remove="settings\resourceorgs.json" />
-    <Content Remove="settings\resourceprojappsvcs.json" />
-    <Content Remove="settings\resourcetypes.json" />
-    <Content Remove="settings\resourceunitdepts.json" />
-    <Content Remove="settings\resourcevmroles.json" />
-    <Content Remove="settings\resoureceunitdepts.json" />
-    <Content Remove="Shared\Admin.razor" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="wwwroot\Screenshots\AdminPasswordPrompt.png" />
-    <None Include="wwwroot\Screenshots\Containers1.png" />
-    <None Include="wwwroot\Screenshots\ConfigurationPageAdmin.png" />
-    <None Include="wwwroot\Screenshots\Containers2.png" />
-    <None Include="wwwroot\Screenshots\FileShare.png" />
-    <None Include="wwwroot\Screenshots\HomePage.png" />
-    <None Include="wwwroot\Screenshots\MountStorage.png" />
-    <None Include="wwwroot\Screenshots\ReferencePage.png" />
+    <None Remove="Components\General\NavMenu.razor.css" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="BlazorDownloadFile" Version="2.4.0.2" />
-    <PackageReference Include="Blazored.Modal" Version="7.1.0" />
-    <PackageReference Include="Blazored.Toast" Version="4.1.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.19.5" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.5.0" />
-    <PackageReference Include="System.Collections" Version="4.3.0" />
-    <PackageReference Include="System.Runtime.Caching" Version="7.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Update="wwwroot\images\runnoft.png">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Update="wwwroot\js\jquery-3.7.0.min.js">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </Content>
+    <PackageReference Include="Blazored.Modal" Version="7.3.1" />
+    <PackageReference Include="Blazored.Toast" Version="4.2.1" />
+    <PackageReference Include="Markdig" Version="0.37.0" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.20.1" />
+    <PackageReference Include="PSC.Blazor.Components.MarkdownEditor" Version="8.0.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore.Swagger" Version="6.6.2" />
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.6.2" />
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.6.2" />
+    <PackageReference Include="System.Runtime.Caching" Version="8.0.0" />
   </ItemGroup>
 
   <ItemGroup>
