@@ -11,8 +11,16 @@ using ResourceType = AzureNamingTool.Models.ResourceType;
 
 namespace AzureNamingTool.Services
 {
+    /// <summary>
+    /// Service for managing Resource Types
+    /// </summary>
     public class ResourceTypeService
     {
+        /// <summary>
+        /// Retrieves a list of resource types.
+        /// </summary>
+        /// <param name="admin">Flag indicating whether to include all resource types or only enabled ones for admin.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response contains the list of resource types.</returns>
         public static async Task<ServiceResponse> GetItems(bool admin = true)
         {
             ServiceResponse serviceResponse = new();
@@ -46,6 +54,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Retrieves a resource type by its ID.
+        /// </summary>
+        /// <param name="id">The ID of the resource type.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response contains the resource type.</returns>
         public static async Task<ServiceResponse> GetItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -80,6 +93,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Adds or updates a resource type.
+        /// </summary>
+        /// <param name="item">The resource type to add or update.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response contains the result of the operation.</returns>
         public static async Task<ServiceResponse> PostItem(ResourceType item)
         {
             ServiceResponse serviceResponse = new();
@@ -92,9 +110,6 @@ namespace AzureNamingTool.Services
                     serviceResponse.ResponseObject = "Short name must be alphanumeric.";
                     return serviceResponse;
                 }
-
-                // Force lowercase on the shortname
-                item.ShortName = item.ShortName.ToLower();
 
                 // Get list of items
                 var items = await ConfigurationHelper.GetList<ResourceType>();
@@ -131,7 +146,7 @@ namespace AzureNamingTool.Services
                     }
 
                     // Write items to file
-                    await ConfigurationHelper.WriteList<ResourceType>(items.OrderBy(x => x.Id).ToList());
+                    await ConfigurationHelper.WriteList<ResourceType>([.. items.OrderBy(x => x.Id)]);
                     serviceResponse.ResponseObject = "Resource Type added/updated!";
                     serviceResponse.Success = true;
                 }
@@ -149,6 +164,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Deletes a resource type by its ID.
+        /// </summary>
+        /// <param name="id">The ID of the resource type to delete.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response indicates the success of the operation.</returns>
         public static async Task<ServiceResponse> DeleteItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -188,6 +208,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Adds or updates a resource type.
+        /// </summary>
+        /// <param name="items">The resource type to add or update.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response contains the result of the operation.</returns>
         public static async Task<ServiceResponse> PostConfig(List<ResourceType> items)
         {
             ServiceResponse serviceResponse = new();
@@ -200,8 +225,6 @@ namespace AzureNamingTool.Services
                 // Determine new item id
                 foreach (ResourceType item in items)
                 {
-                    // Force lowercase on the shortname
-                    item.ShortName = item.ShortName.ToLower();
                     item.Id = i;
                     newitems.Add(item);
                     i += 1;
@@ -220,10 +243,15 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Retrieves the categories of resource types.
+        /// </summary>
+        /// <param name="types">The list of resource types.</param>
+        /// <returns>A list of resource type categories.</returns>
         public static List<string> GetTypeCategories(List<ResourceType> types)
         {
             ServiceResponse serviceResponse = new();
-            List<string> categories = new();
+            List<string> categories = [];
 
             foreach (ResourceType type in types)
             {
@@ -233,7 +261,7 @@ namespace AzureNamingTool.Services
                 {
                     if (category.Contains('/'))
                     {
-                        category = category[..category.IndexOf("/")];
+                        category = category[..category.IndexOf('/')];
                     }
 
                     if (!categories.Contains(category))
@@ -250,14 +278,20 @@ namespace AzureNamingTool.Services
             return categories;
         }
 
+        /// <summary>
+        /// Retrieves a filtered list of resource types based on the specified filter.
+        /// </summary>
+        /// <param name="types">The list of resource types.</param>
+        /// <param name="filter">The filter to apply.</param>
+        /// <returns>The filtered list of resource types.</returns>
         public static List<ResourceType> GetFilteredResourceTypes(List<ResourceType> types, string filter)
         {
             ServiceResponse serviceResponse = new();
-            List<ResourceType> currenttypes = new();
+            List<ResourceType> currenttypes = [];
             // Filter out resource types that should have name generation
             if (!String.IsNullOrEmpty(filter))
             {
-                currenttypes = types.Where(x => x.Resource.ToLower().StartsWith(filter.ToLower() + "/") && x.Property.ToLower() != "display name" && !String.IsNullOrEmpty(x.ShortName)).ToList();
+                currenttypes = types.Where(x => x.Resource.StartsWith(filter + "/", StringComparison.CurrentCultureIgnoreCase) && !x.Property.Equals("display name", StringComparison.CurrentCultureIgnoreCase) && !String.IsNullOrEmpty(x.ShortName)).ToList();
             }
             else
             {
@@ -266,6 +300,11 @@ namespace AzureNamingTool.Services
             return currenttypes;
         }
 
+        /// <summary>
+        /// Refreshes the resource types by retrieving the latest data from the repository.
+        /// </summary>
+        /// <param name="shortNameReset">Flag indicating whether to reset the short names of existing resource types.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response indicates the success of the operation.</returns>
         public static async Task<ServiceResponse> RefreshResourceTypes(bool shortNameReset = false)
         {
             ServiceResponse serviceResponse = new();
@@ -283,7 +322,7 @@ namespace AzureNamingTool.Services
                         if (!String.IsNullOrEmpty(refreshdata))
                         {
                             var newtypes = new List<ResourceType>();
-                            var options = new JsonSerializerOptions
+                            JsonSerializerOptions options = new()
                             {
                                 PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
                                 PropertyNameCaseInsensitive = true
@@ -363,6 +402,12 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Updates the components of a resource type based on the specified operation and component ID.
+        /// </summary>
+        /// <param name="operation">The operation to perform on the resource type components.</param>
+        /// <param name="componentid">The ID of the component to update.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response indicates the success of the operation.</returns>
         public static async Task<ServiceResponse> UpdateTypeComponents(string operation, int componentid)
         {
             ServiceResponse serviceResponse = new();
@@ -383,7 +428,7 @@ namespace AzureNamingTool.Services
                                 List<ResourceType> resourceTypes = (List<ResourceType>)serviceResponse.ResponseObject!;
                                 if (GeneralHelper.IsNotNull(resourceTypes))
                                 {
-                                    List<string> currentvalues = new();
+                                    List<string> currentvalues = [];
                                     // Update all the resource type component settings
                                     foreach (ResourceType currenttype in resourceTypes)
                                     {
@@ -394,7 +439,7 @@ namespace AzureNamingTool.Services
                                                 if (!currentvalues.Contains(component))
                                                 {
                                                     currentvalues.Add(component);
-                                                    currenttype.Optional = String.Join(",", currentvalues.ToArray());
+                                                    currenttype.Optional = String.Join(",", [.. currentvalues]);
                                                     await ResourceTypeService.PostItem(currenttype);
                                                 }
                                                 break;
@@ -403,7 +448,7 @@ namespace AzureNamingTool.Services
                                                 if (currentvalues.Contains(component))
                                                 {
                                                     currentvalues.Remove(component);
-                                                    currenttype.Optional = String.Join(",", currentvalues.ToArray());
+                                                    currenttype.Optional = String.Join(",", [.. currentvalues]);
                                                     await ResourceTypeService.PostItem(currenttype);
                                                 }
                                                 break;
@@ -412,7 +457,7 @@ namespace AzureNamingTool.Services
                                                 if (!currentvalues.Contains(component))
                                                 {
                                                     currentvalues.Add(component);
-                                                    currenttype.Exclude = String.Join(",", currentvalues.ToArray());
+                                                    currenttype.Exclude = String.Join(",", [.. currentvalues]);
                                                     await ResourceTypeService.PostItem(currenttype);
                                                 }
                                                 break;
@@ -421,7 +466,7 @@ namespace AzureNamingTool.Services
                                                 if (currentvalues.Contains(component))
                                                 {
                                                     currentvalues.Remove(component);
-                                                    currenttype.Exclude = String.Join(",", currentvalues.ToArray());
+                                                    currenttype.Exclude = String.Join(",", [.. currentvalues]);
                                                     await ResourceTypeService.PostItem(currenttype);
                                                 }
                                                 break;
@@ -452,6 +497,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Validates the name of a resource type.
+        /// </summary>
+        /// <param name="validateNameRequest">The validate name request containing the resource type information.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation. The service response contains the validation result.</returns>
         public static async Task<ServiceResponse> ValidateResourceTypeName(ValidateNameRequest validateNameRequest)
         {
             ServiceResponse serviceResponse = new();
@@ -485,7 +535,7 @@ namespace AzureNamingTool.Services
                         List<ResourceType> resourceTypes = (List<ResourceType>)serviceResponse.ResponseObject!;
                         if (GeneralHelper.IsNotNull(resourceTypes))
                         {
-                            ResourceType resourceType = null;
+                            ResourceType? resourceType = null;
                             if (GeneralHelper.IsNotNull(validateNameRequest.ResourceTypeId))
                             {
                                 // Get the specified resoure type by id
